requirements.txt
git+https://github.com/Rapptz/discord.py
aiohttp
asyncio
jishaku
python-dotenv
bs4
git+https://github.com/Rapptz/discord-ext-menus